Run specs in a random order
**Why**: Specs should not be dependent on any particular order. We
want to surface any order dependencies as bugs.
