+++ app/jobs/end_product_sync_job.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../exceptions/bgs_sync_error"
-
 # This job syncs an EndProductEstablishment (end product manager) with up to date BGS and VBMS data
 class EndProductSyncJob < CaseflowJob
   queue_as :low_priority
@@ -12,7 +10,7 @@ class EndProductSyncJob < CaseflowJob
 
     begin
       EndProductEstablishment.find(end_product_establishment_id).sync!
-    rescue ::TransientBGSSyncError => error
+    rescue BGSError::TransientBGSSyncError => error
       # we don't care about transient errors in Sentry since it will alert us. we'll just try again later.
       Rails.logger.error error
     rescue StandardError => error
