+++ spec/feature/reader/reader_spec.rb
@@ -27,10 +27,6 @@ def scrolled_amount(child_class_name)
   EOS
 end
 
-def scroll_element_to_view(element)
-  page.execute_script("document.getElementById('#{element}').scrollIntoView()")
-end
-
 def scroll_to_bottom(id: nil, class_name: nil)
   page.driver.evaluate_script <<-EOS
     function() {
